[REFACTOR] ExportModal hooks 이동
- component 코드인 `ExportModal`에 있는 로직을 `useExportStyle`로 이동
- `getStringifyStyleObject`에서 markers length/nullish 판단을 nullish 병합 연산으로 처리
